server/requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
ddgs
pdfplumber
python-docx
python-dotenv

